[REVIEW] Updating rmm header, fixing e2e tests with deprecated drop_column function and fixing ORC args changes

+++ engine/src/io/data_parser/ParquetParser.h
@@ -26,7 +26,7 @@ public:
 	std::unique_ptr<ral::frame::BlazingTable> parse_batch(
 		std::shared_ptr<arrow::io::RandomAccessFile> file,
 		const Schema & schema,
-		std::vector<size_t> column_indices,
+		std::vector<int> column_indices,
 		std::vector<cudf::size_type> row_groups);
 
 	void parse_schema(std::shared_ptr<arrow::io::RandomAccessFile> file, Schema & schema);
